Fix type from "float" -> "double"

+++ include/Lightracer/RacerTrackCollisionResolver.hpp
@@ -13,7 +13,7 @@ private:
    Racer &player;
    Track *track;
 
-   float &OMG_DeltaTime;
+   double &OMG_DeltaTime;
    int &index_of_last_track_segment_that_collides;
    ALLEGRO_SAMPLE_INSTANCE *passthough_sample_instance;
    int &num_laps_to_win;
@@ -25,7 +25,7 @@ public:
    RacerTrackCollisionResolver(
         Racer *racer
       , Track *track
-      , float &OMG_DeltaTime
+      , double &OMG_DeltaTime
       , int &index_of_last_track_segment_that_collides
       , ALLEGRO_SAMPLE_INSTANCE *passthough_sample_instance
       , int &num_laps_to_win
